% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_panels.R
\name{add_panels}
\alias{add_panels}
\title{add panels to the regression table}
\usage{
add_panels(
  regression_table,
  num_panels,
  bold = FALSE,
  italic = TRUE,
  latex_gap_space = "0.5cm",
  panel_labels,
  rows_per_model,
  hline_after = FALSE
)
}
\arguments{
\item{regression_table}{the regression table to be given panels to.}

\item{num_panels}{the number of panels in the regression table. Passed in from panelsummary.}

\item{bold}{whether the panel labels should be bold or not.}

\item{italic}{whether the panel labels should be in italics.}

\item{latex_gap_space}{the amount of space between the two panels. This is the optimal spacing.}

\item{panel_labels}{the text to come after each Panel A:... and Panel B:...}

\item{rows_per_model}{the number of rows in each of the panels.}

\item{hline_after}{whether to add a horizontal line after each Panel}
}
\description{
Adds panels (such as Panel A, Panel B,...) to the regression tables.
}
\keyword{internal}
