% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_means.R
\name{change_to_lists}
\alias{change_to_lists}
\title{Changes all models passed into panelsummary into a list}
\usage{
change_to_lists(models)
}
\arguments{
\item{models}{Input list. This is the input list of models taken from the user.}
}
\value{
A list of lists. It will change any non-list component to a list.
}
\description{
Changes all models passed into panelsummary into a list
}
\keyword{This}
\keyword{`change_to_lists`}
\keyword{a}
\keyword{all}
\keyword{applying}
\keyword{changes}
\keyword{correctly}
\keyword{done}
\keyword{ensure}
\keyword{functions}
\keyword{getting}
\keyword{internal}
\keyword{into}
\keyword{is}
\keyword{list.}
\keyword{lists.}
\keyword{means}
\keyword{models}
\keyword{mostly}
\keyword{over}
\keyword{panelsummary::panelsummary}
\keyword{passed}
\keyword{that}
\keyword{the}
\keyword{to}
\keyword{when}
