% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_means.R
\name{check_class_fixest}
\alias{check_class_fixest}
\title{Checks that the class of each regression is of the type "fixest"}
\usage{
check_class_fixest(models)
}
\arguments{
\item{models}{Input list. This is the input list of models taken from the user.}
}
\value{
A Boolean of TRUE/FALSE indicating whether the elements are of fixest type.
}
\description{
`check_class_fixest` ensures that each class is of the type "fixest" so that the mean of the
dependent variable can be taken using fixest built-in methods.
}
\keyword{internal}
