% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dimensions_match.R
\name{check_dimensions_match}
\alias{check_dimensions_match}
\title{check if the dimensions work}
\usage{
check_dimensions_match(models, num_panels)
}
\arguments{
\item{models}{the regression models to check the dimensions of}

\item{num_panels}{the number of panels passed from panelsummary}
}
\description{
check if the dimensions work
}
\keyword{internal}
