% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_means.R
\name{connect_means}
\alias{connect_means}
\title{Merges the means to the modelsummary output dataframe}
\usage{
connect_means(panel_df, means)
}
\arguments{
\item{panel_df}{The modelsummary dataframe supplied from modelsummary::modelsummary(x, output = "data.frame")}

\item{means}{The list of named vectors of means which correspond to each model.}
}
\value{
A data.frame with the attached means.
}
\description{
`connect_means` connects the means to the modelsummary dataframe.
}
