% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{create_alignment}
\alias{create_alignment}
\title{creates alignment of left, center, center, ...}
\usage{
create_alignment(number_models)
}
\description{
creates alignment of left, center, center, ...
}
\keyword{internal}
