% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{create_column_names}
\alias{create_column_names}
\title{creates the column names of (1), (2), ...}
\usage{
create_column_names(number_models)
}
\description{
creates the column names of (1), (2), ...
}
\keyword{internal}
