% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_means.R
\name{create_mean_fixest}
\alias{create_mean_fixest}
\title{creates a custom glance function for fixest objects with a mean}
\usage{
create_mean_fixest()
}
\description{
creates a custom glance function for fixest objects with a mean
}
\keyword{internal}
