% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_means.R
\name{create_mean_lm}
\alias{create_mean_lm}
\title{creates a custom glance function for lm objects with a mean}
\usage{
create_mean_lm()
}
\description{
creates a custom glance function for lm objects with a mean
}
\keyword{internal}
