% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_panels_null.R
\name{create_panels_null}
\alias{create_panels_null}
\title{creates panel labels when the argument is null}
\usage{
create_panels_null(num_panels)
}
\arguments{
\item{num_panels}{the number of panels in the regression table.}
}
\value{
default panel labels named "Panel A:", "Panel B:", ..."Panel E:"
}
\description{
creates panel labels when the argument is null
}
\keyword{internal}
