% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{create_pretty_numbers}
\alias{create_pretty_numbers}
\title{creates prettyNum with commas for anything larger than 1000}
\usage{
create_pretty_numbers(df)
}
\description{
creates prettyNum with commas for anything larger than 1000
}
\keyword{internal}
