% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_lpanel_fe}
\alias{get_lpanel_fe}
\title{gets the number of fixed effects in the last panel for collapse_fe purposes}
\usage{
get_lpanel_fe(df, panels)
}
\description{
gets the number of fixed effects in the last panel for collapse_fe purposes
}
\keyword{internal}
