% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_means.R
\name{get_means_fixest}
\alias{get_means_fixest}
\title{Generates the mean of the dependent variable for each model}
\usage{
get_means_fixest(models, fmt)
}
\arguments{
\item{models}{A list of lists. Each list should contain fixest objects.}

\item{fmt}{An integer. This will denote how many decimal places to round to for the table.}
}
\value{
A list of named vectors denoting the means of each model.
}
\description{
`get_means_fixest` generates the mean of the dependent variable for each model supplied to the panelsummary::panelsummary.
This function only accepts fixest objects as it uses built-in fixest fitstat to compute the means.
}
\keyword{internal}
