% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_panel_indices}
\alias{get_panel_indices}
\title{gets the row numbers for break points in the panels}
\usage{
get_panel_indices(panel)
}
\description{
gets the row numbers for break points in the panels
}
\keyword{internal}
