% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_panel_indices_collapse}
\alias{get_panel_indices_collapse}
\title{gets the indices of each of the breaks in the panel}
\usage{
get_panel_indices_collapse(panel_df, num_panels)
}
\description{
gets the indices of each of the breaks in the panel
}
\keyword{internal}
