% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models_supported.R
\name{models_supported}
\alias{models_supported}
\title{all models supported by panelsummary}
\usage{
models_supported()
}
\value{
a list of all modeltypes supported by panelsummary
}
\description{
all models supported by panelsummary
}
\examples{

models_supported()

}
