% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelsummary-package.R
\docType{package}
\name{panelsummary-package}
\alias{panelsummary-package}
\alias{_PACKAGE}
\title{panelsummary: Create Publication-Ready Regression Tables with Panels}
\description{
Create an automated regression table that is well-suited for models that are estimated with multiple dependent variables. 'panelsummary' extends 'modelsummary' (Arel-Bundock, V. (2022) \doi{10.18637/jss.v103.i01}) by allowing regression tables to be split into multiple sections with a simple function call. Utilize familiar arguments such as fmt, estimate, statistic, vcov, conf_level, stars, coef_map, coef_omit, coef_rename, gof_map, and gof_omit from 'modelsummary' to clean the table, and additionally, add a row for the mean of the dependent variable without external manipulation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/michaeltopper1/panelsummary}
  \item \url{https://michaeltopper1.github.io/panelsummary/}
  \item Report bugs at \url{https://github.com/michaeltopper1/panelsummary/issues}
}

}
\author{
\strong{Maintainer}: Michael Topper \email{miketopper123@gmail.com}

Other contributors:
\itemize{
  \item Vincent Arel-Bundock (Some documentation is inherited from 'modelsummary') [copyright holder]
}

}
\keyword{internal}
