% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{remove_fe}
\alias{remove_fe}
\title{removes all fixed effects except the final panel (for collapse_fe)}
\usage{
remove_fe(panel_df, num_panels)
}
\description{
removes all fixed effects except the final panel (for collapse_fe)
}
\keyword{internal}
