% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{shift_means}
\alias{shift_means}
\title{shifts the custom glance means to above observations and renames}
\usage{
shift_means(df)
}
\description{
shifts the custom glance means to above observations and renames
}
\keyword{internal}
