% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dimensions_match.R
\name{check_dimensions_match}
\alias{check_dimensions_match}
\title{check if the dimensions work}
\usage{
check_dimensions_match(models, panel_labels)
}
\arguments{
\item{models}{the regression models to check the dimensions of}

\item{panel_labels}{the labels for each panel.}
}
\description{
check if the dimensions work
}
\keyword{internal}
