% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panelsummary-package.R
\docType{package}
\name{panelsummary-package}
\alias{panelsummary-package}
\alias{_PACKAGE}
\title{panelsummary: Create Publication-Ready Regression Tables with Panels}
\description{
Creates publication-ready regression tables with panels. Solves the hassle of adding multiple panels (e.g., Panel A/Panel B) to a single regression table.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/michaeltopper1/panelsummary}
  \item \url{https://michaeltopper1.github.io/panelsummary/}
  \item Report bugs at \url{https://github.com/michaeltopper1/panelsummary/issues}
}

}
\author{
\strong{Maintainer}: Michael Topper \email{miketopper123@gmail.com}

Other contributors:
\itemize{
  \item Vincent Arel-Bundock (Some documentation is inherited from modelsummary) [copyright holder]
}

}
\keyword{internal}
